tests for HeaderView and BlockView, fixed #144
