+++ ethcore/src/views/mod.rs
@@ -1,5 +1,25 @@
-mod transaction;
-mod views;
+// Copyright 2015, 2016 Ethcore (UK) Ltd.
+// This file is part of Parity.
 
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+//! Block oriented views onto rlp.
+
+mod block;
+mod header;
+mod transaction;
+
+pub use self::block::BlockView;
+pub use self::header::HeaderView;
 pub use self::transaction::TransactionView;
-pub use self::views::*;
